+++ app/main.js
@@ -1,0 +1,7 @@
+import darkMode from "./darkMode/darkModeToggle.js";
+
+darkMode();
+
+import initSlideShow from "./slideShow/slideShow.js";
+
+initSlideShow(0);
