+++ app/slideShow/utils/globalVars.js
@@ -1,0 +1,12 @@
+const globalData = {};
+
+const globalVar = {
+  get(key) {
+    return globalData[key];
+  },
+  set(key, data) {
+    globalData[key] = data;
+  },
+};
+
+export default globalVar;
