+++ assets/slideShowData/projects.JSON
@@ -1,0 +1,20 @@
+[
+  {
+    "projImg": "assets/photos/project1.jpg",
+    "projTitle": "0. Hack.api",
+    "projDescription": "small text, this is a small description",
+    "projLink": "project1.html"
+  },
+  {
+    "projImg": "assets/photos/project2.jpg",
+    "projTitle": "1. Space Image Classification",
+    "projDescription": "150 words, Project Description. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce sodales, dui aliquam venenatis finibus, metus velit aliquam augue, at consequat eros sapien ut mi. Pellentesque congue justo nec ante egestas, vel vulputate ex tempus. Quisque at leo varius, volutpat mi sit amet, laoreet turpis. Proin elementum faucibus purus eget pretium. Aliquam sit amet sem mattis, gravida tellus nec, mollis est. Fusce non velit id justo posuere ultrices. Ut finibus tempus neque, in interdum ligula scelerisque sit amet. Quisque bibendum tincidunt mi, sit amet faucibus urna dapibus non. Fusce eget dictum mauris, sit amet consectetur leo. Nunc enim lectus, iaculis sed justo sit amet, porttitor tempus libero,turpis. Proin elementum faucibus purus eget pretium. Aliquam sit amet sem mattis, gravida tellus nec, mollis est. Fusce non velit id justo posuere ultrices.tempus libero,turpis. Proin elementum faucibus purus eget pretium. Aliquam sit amet sem mattis, gravida tellus nec, mollis est. Fusce non velit id justo posuere ultrices.",
+    "projLink": "project1.html"
+  },
+  {
+    "projImg": "assets/photos/project3.jpg",
+    "projTitle": "2. kek Project",
+    "projDescription": "Project Description. Lorem ipsum dolor sit amet, consectetur adipiscing elit. Fusce sodales, dui aliquam venenatis finibus, metus velit aliquam augue, at consequat eros sapien ut mi. Pellentesque congue justo nec ante egestas, vel vulputate ex tempus. Quisque at leo varius, volutpat mi sit amet, laoreet turpis. Proin elementum faucibus purus eget pretium. Aliquam sit amet sem mattis, gravida tellus nec, mollis est. Fusce non velit id justo posuere ultrices. Ut finibus tempus neque, in interdum ligula scelerisque sit amet. Quisque bibendum tincidunt mi, sit amet faucibus urna dapibus non. Fusce eget dictum mauris, sit amet consectetur leo. Nunc enim lectus, iaculis sed justo sit amet, porttitor tempus libero",
+    "projLink": "project1.html"
+  }
+]
