+++ index.html
@@ -85,7 +85,6 @@
       </div>
     </footer>
 
-    <script src="darkModeToggle.js"></script>
-    <script src="slideShow.js"></script>
+    <script src="./app/main.js" type="module"></script>
   </body>
 </html>
